chore(geo-ip): convert geo-ip to store
